Improved vertical movements with wrapped lines
After turning on soft wrapping of lines, vertical movements can
sometimes be jarring and annoying when editing files with many visual
lines, such as LaTeX, Markdown, or text files.

We rebind 'j' and 'k' to use 'gj' and 'gk' whenever we move vertically.
We retain the normal behavior when using '3j' or '6k' to keep relative
line jumping consistent.