Reintroduce `nvim-treesitter` on without blocking UI
Previously, when updating to the `nvim-treesitter` main branch, loading
parsers by `filetype` would block the UI from loading, and thus cause
noticeable lag when entering a file. We nuked `treesitter` in a previous
commit to get rid of this lag.

We are now reintroducing `nvim-treesitter` back into the configuration,
and enable Treesitter highlights after a `defer_fn` of around 20ms, to
allow the UI to load first. This mitigates most of the perceivable input
lag.